# docs/openapi.yaml
openapi: 3.1.0
info:
  title: piper API
  version: 0.1.0
  description: |
    Project-owned resources are always exposed below `/api/projects/{project_id}`.
    Collection POST operations return `201 Created`, successful mutations with no
    response body return `204 No Content`, JSON field names use snake_case, and
    errors use the shape `{"error":"message"}`.
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
paths:
...
  /api/system/storage/settings/test:
    post:
      description: System admin only. Checks a candidate configuration without persisting it.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/StorageConfig" }
      responses:
        "200":
          description: Reachability result.
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  message: { type: string }
...
  /api/projects/{project_id}:
    parameters:
      - $ref: "#/components/parameters/ProjectID"
    get:
      responses:
        "200":
          description: Project.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Project" }
    delete:
      description: System admin only. Stops the project's live runs, notebooks, and services before deleting it.
      responses:
        "204":
          description: Deleted.
        "409":
          description: The last remaining project cannot be deleted.
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    ProjectID:
      name: project_id
      in: path
      required: true
      schema: { type: string }
...
  schemas:
...
    Project:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        description: { type: string }
        owner_member_id: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    StorageConfig:
      type: object
      properties:
        url: { type: string }
        disabled: { type: boolean }
        token: { type: string }
        credentialRef: { type: string }
...